+++ eventservice/api/controllers/events.py
@@ -37,7 +37,8 @@ from controllers.images.s3 import S3EventsImages
 from controllers.images import localfile
 
 from utils.cache import memoize , memoize_query, CACHE_GET_EVENTS, CACHE_GET_EVENT, CACHE_GET_EVENTIMAGES, CACHE_GET_CATEGORIES
-from utils.group_auth import get_group_ids, get_group_memberships, check_group_event_admin_access, check_permission_access_event
+from utils.group_auth import get_group_ids, get_group_memberships, check_group_event_admin_access, check_permission_access_event, \
+    check_all_group_event_admin
 
 logging.Formatter.converter = gmtime
 logging.basicConfig(level=logging.INFO, datefmt='%Y-%m-%dT%H:%M:%S',
@@ -48,8 +49,21 @@ __logger = logging.getLogger("events_building_block")
 def search():
     group_ids = list()
     include_private_events = False
+
+    is_all_group_event = False
+
+    # check permission if the user has access to all group events.
+    try:
+        is_all_group_event = check_all_group_event_admin()
+    except Exception as ex:
+        msg = "Failed to parse the id token."
+        __logger.exception(msg, ex)
+        abort(500)
+
     try:
         include_private_events, group_ids = get_group_ids()
+        # check if the user is in all group then give some boolean checking
+        # use the field called createdByGroupId to check if it is group event
     except Exception as ex:
         __logger.exception(msgs.ERR_MSG_GET_GROUP)
         abort(500)
@@ -57,7 +71,11 @@ def search():
     args = request.args
     query = dict()
     try:
-        query = query_params.format_query(args, query, include_private_events, group_ids)
+        # if all group group then give the query with all the group event
+        if is_all_group_event:
+            query = query_params.format_query(args, query, True, None, True)
+        else:
+            query = query_params.format_query(args, query, include_private_events, group_ids)
     except Exception as ex:
         __logger.exception(ex)
         abort(400)
@@ -396,7 +414,20 @@ def patch(event_id):
 
 
 def delete(event_id):
-    auth_middleware.authorize(auth_middleware.ROKWIRE_EVENT_WRITE_GROUPS)
+    can_delete = False
+
+    # check permission if the user has access to all group events.
+    try:
+        is_all_group_event = check_all_group_event_admin()
+        if is_all_group_event:
+            can_delete = True
+    except Exception as ex:
+        msg = "Failed to parse the id token."
+        __logger.exception(msg, ex)
+        abort(500)
+
+    if not can_delete:
+        auth_middleware.authorize(auth_middleware.ROKWIRE_EVENT_WRITE_GROUPS)
 
     if not ObjectId.is_valid(event_id):
         abort(400)
@@ -417,8 +448,9 @@ def delete(event_id):
         __logger.exception(msgs.ERR_MSG_GET_EVENT)
         abort(500)
 
-    if not check_group_event_admin_access(event, group_memberships):
-        abort(401)
+    if not can_delete:
+        if not check_group_event_admin_access(event, group_memberships):
+            abort(401)
 
     try:
         db = get_db()